Fixing track issues, aligning with FIXM per FAA suggestion
Removing track_magnetic_north_deg and track_true_north_deg and replacing with track_bearing and track_bearing_reference.  The former is the degree value for direction of travel and the second is the reference (true north or magnetic north) for that direction.  This aligns to the FIXM definitions and allows for easier translation.  Also removes potential conflicts when a USS/Operator provides both track_magnetic_north_deg and track_true_north_deg which may have opposing values.