Made Metadata required, and made test_identifiers required per #10
We had issues with partners neglecting metadata because it was optional in the model, even though the testing documentation made it required.  It is optional in the API since it will not be required in a future operational system.  To reduce confusion within the testing environment, making metadata required.  Will have to make it optional in the future after TCL4.