#95 Added 403's for endpoints. 404 for operation/{gufi}
We should be returning 403's when the token is a real, valid token but it is insufficeint for the current request.  The most common cause would be the wrong scope presented.  In the future if we have 'aud' claim, then an incorrect aud value would be another good reason for a 403.