#138. Adding initial draft for /enhanced/operations/{gufi}
Will start discussion against this concrete implementation of the API.  Expect that feedback may aid in refining.  Targeting Sprint 4.

Note that this approach was suggested in #87 .  The thinking is that it is easier to have a separate endpoint than to have rules for models excessively dependent upon the access_token and data model.